+++ changes.md
@@ -4,6 +4,7 @@
 
 ## 3.2.0
 * Added assembly build to shieldwall straight.
+* Fix ShieldShape bug where the base height would break in a way that made the shape not actually symetrical.
 
 ## 3.1.0
 * Added shieldwall Set